+++ elastic-job-core/src/test/java/com/dangdang/ddframe/job/plugin/job/type/fixture/FooUnstreamingBatchThroughputDataFlowElasticJob.java
@@ -38,8 +38,11 @@ public final class FooUnstreamingBatchThroughputDataFlowElasticJob extends Abstr
     public int processData(final JobExecutionMultipleShardingContext shardingContext, final List<Object> data) {
         int result = 0;
         for (Object each : data) {
-            if (jobCaller.processData(each)) {
-                result++;
+            try {
+                if (jobCaller.processData(each)) {
+                    result++;
+                }
+            } catch (final NullPointerException ex) {
             }
         }
         return result;